Pass resolving to component when rendering
At the moment the only way to get access to the resolving variable
is to replace the `Component` prop with a `render` prop. This can be
a lot of overhead for a simple route, so passing this variable in as
a prop to the component should make the API easier to use and more
consistent with what render receives.